use rewards api endpoint
